♻️ rename modules
