+++ src/vector_map.py
@@ -6,11 +6,11 @@ import numpy
 from shapely import geometry, ops
 
 # from PIL import Image, ImageDraw, ImageFilter
-from . import O4_OSM_Utils as osm
-from . import O4_UI_Utils as ui
-from . import O4_Vector_Utils as vect
 from . import airport as APT
 from . import dem, filenames, geo
+from . import osm as osm
+from . import ui as ui
+from . import vector as vect
 
 good_imagery_list = ()
